refactor(LegMatcher): Extract LegMatcher class.
